+++ lib/serverpod_vps.dart
@@ -1,1 +1,2 @@
+export 'src/docker_network_name.dart';
 export 'src/template_generator.dart';